+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Field `sqs_endpoint` added to the `s3` input.
+- Operator `delete` added to `metadata` processor.
 
 ## 3.2.0 - 2019-09-27